refactor : 승자를 출력하는 부분에서 객체에게 요청하는 코드로 수정
