+++ src/main/java/racingGame/controller/RacingGame.java
@@ -10,34 +10,16 @@ import racingGame.view.InputView;
 import racingGame.view.OutputView;
 
 public class RacingGame {
-    private int maxPosition;
-
     public void run() {
-        int movementNumber;
-
+        int movementNumber = InputView.getMovementNumber();
         Cars cars = new Cars(CarFactory.createByNames(InputView.getCarName()));
-        movementNumber = InputView.getMovementNumber();
 
         for (int i = 0; i < movementNumber; i++) {
             cars.moveCars();
             OutputView.printCars(cars);
         }
-        maxPosition = cars.getMaximumPosition();
-        OutputView.printWinners(getWinners(cars));
+        OutputView.printWinners(cars.getWinnersList());
     }
 
-    public List<Car> getWinners(final Cars cars) {
-        List<Car> winners = new ArrayList<>();
 
-        for (Car car : cars.getCars()) {
-            addIfWinner(winners, car);
-        }
-        return winners;
-    }
-
-    private void addIfWinner(List<Car> winners, final Car car) {
-        if (car.isWinnerPosition(maxPosition)) {
-            winners.add(car);
-        }
-    }
 }
